+++ webapp/src/main/java/io/camunda/operate/webapp/management/dto/GetBackupStateResponseDto.java
@@ -11,7 +11,7 @@ import java.util.Objects;
 
 public class GetBackupStateResponseDto {
 
-  private Integer backupId;
+  private Long backupId;
   private BackupStateDto state;
   private String failureReason;
   private List<GetBackupStateResponseDetailDto> details;
@@ -19,15 +19,15 @@ public class GetBackupStateResponseDto {
   public GetBackupStateResponseDto() {
   }
 
-  public GetBackupStateResponseDto(Integer backupId) {
+  public GetBackupStateResponseDto(Long backupId) {
     this.backupId = backupId;
   }
 
-  public Integer getBackupId() {
+  public Long getBackupId() {
     return backupId;
   }
 
-  public GetBackupStateResponseDto setBackupId(Integer backupId) {
+  public GetBackupStateResponseDto setBackupId(Long backupId) {
     this.backupId = backupId;
     return this;
   }
